# Write length and payload separately for double the throughput