@puellanivis 2020-10-27
I think what I would more ideally like to see, is things to marshal into a slice that has space for the length already in it, but just unset. Then before writing it, the length is then written into that header, and then written as a whole.

The `sync.Pool` work being done here probably isn’t the correct solution either. Returning a pointer to an array is really weird when the array is already has pointer semantics. And in fact, on a 64-bit machine, we end up with just as much space using the pointer to an array as we would using `make([]byte, 4)`:
* `unsafe.Sizeof(uintptr) /* pointer to */ + unsafe.Sizeof(uintptr) + unsafe.Sizeof(int) /* array header */ + 4*unsafe.Sizeof(byte)  = 28`
* `unsafe.Sizeof(reflect.SliceHeader) + 4*unsafe.Sizeof(byte) = 28`

I think that probably everything here would still continue to work if we break up all of these writes, I have been bitten before by weird segmentation issues where `Write`s are not being done as an whole together. There’s likely already still buffers pulling the various writes together into a single transmission, since writing a bunch of half-packets would be quite inefficient… (Nagle’s Algorithm, etc)