@greatroar 2020-10-27
> I think what I would more ideally like to see, is things to marshal into a slice that has space for the length already in it, but just unset. Then before writing it, the length is then written into that header, and then written as a whole.

The code in this PR is not as elegant as it could be, but it's relatively local. I can try to implement a more thorough solution, but it will (1) be a much larger patch, and (2) the packet and the payload really need to be separated to get an O(1) sendPacket rather than the current O(n).

> Returning a pointer to an array is really weird when the array is already has pointer semantics.

I think you're confusing arrays and slices. A [4]byte is just four bytes, without a slice header, and with value semantics. When you assign one to an interface{}, it gets copied to the heap. Try

```go
func BenchmarkArrayPool(b *testing.B) {
	pool := sync.Pool{New: func() interface{} { return [4]byte{} }}

	for i := 0; i < b.N; i++ {
		a := pool.Get().([4]byte)
		pool.Put(a) // array copied to the heap
	}
}

func BenchmarkPointerPool(b *testing.B) {
	pool := sync.Pool{New: func() interface{} { return new([4]byte) }}

	for i := 0; i < b.N; i++ {
		a := pool.Get().(*[4]byte)
		pool.Put(a)
	}
}

func BenchmarkSlicePool(b *testing.B) {
	pool := sync.Pool{New: func() interface{} { return make([]byte, 4) }}

	for i := 0; i < b.N; i++ {
		a := pool.Get().([]byte)
		pool.Put(a) // slice header copied to the heap
	}
}
```

On my machine, that gives

```
BenchmarkArrayPool-8     	28204717	        42.4 ns/op	       4 B/op	       1 allocs/op
BenchmarkPointerPool-8   	69263575	        17.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkSlicePool-8     	19285311	        86.2 ns/op	      32 B/op	       1 allocs/op
```

> I have been bitten before by weird segmentation issues where Writes are not being done as an whole together.

With this library? Because I don't see how the SFTP protocol could be affected.