@puellanivis 2020-10-29
So, I’ve played around with this a bit, and came up with the commits in https://github.com/pkg/sftp/tree/more-optimization (also included some various subtle race conditions being dealt with as well), by just allocating and starting after the length, it’s a lot of lines-of-code-change, but simpler in the end without a need for a `sync.Pool`. (Oh, and I found a few incorrect length calculations, that could have :woman_shrugging: been causing reallocations? Results unclear, go 1.15 at least seems to be allocating with some buffer room sometimes.)

The other improvements were mostly with an eye towards improving the `client.go` code as well… but the results of my first go are… hazy about how good they might be. They’re significantly better under write delay, but only “meh” at best without… (more concurrency == more overhead, but also better responsiveness per channel when slow). And does ~20 kB/op really matter, when both maybe-solutions can get about the same throughput, and we‘re already cutting out 100s of MB/op? :woman_shrugging:

I’ll have something tomorrow either way, either the code sitting uncommitted on my laptop, or a restart that isn’t overhauling `client.go` as much.