@greatroar 2020-10-29
I've benchmarked your branch @ 113f5e9b80e37152e3412ee9a2d6fbd8c61e9b37 against master:

```
name         old time/op    new time/op    delta
Write1k-8       312ms ±12%     311ms ± 6%      ~     (p=0.796 n=10+10)
Write16k-8     56.9ms ± 2%    32.7ms ± 4%   -42.62%  (p=0.000 n=10+10)
Write32k-8     43.2ms ± 3%    17.6ms ± 5%   -59.17%  (p=0.000 n=10+10)
Write128k-8    31.1ms ± 2%    14.7ms ± 8%   -52.77%  (p=0.000 n=10+10)
Write512k-8    17.5ms ± 5%    10.2ms ±10%   -41.66%  (p=0.000 n=10+10)
Write1MiB-8    14.9ms ± 1%     9.2ms ± 3%   -38.19%  (p=0.000 n=8+10)

name         old speed      new speed      delta
Write1k-8    33.7MB/s ±11%  33.8MB/s ± 6%      ~     (p=0.796 n=10+10)
Write16k-8    184MB/s ± 2%   321MB/s ± 4%   +74.35%  (p=0.000 n=10+10)
Write32k-8    243MB/s ± 3%   595MB/s ± 5%  +145.04%  (p=0.000 n=10+10)
Write128k-8   337MB/s ± 2%   715MB/s ± 8%  +112.43%  (p=0.000 n=10+10)
Write512k-8   599MB/s ± 5%  1028MB/s ±10%   +71.65%  (p=0.000 n=10+10)
Write1MiB-8   702MB/s ± 1%  1136MB/s ± 3%   +61.84%  (p=0.000 n=8+10)

name         old alloc/op   new alloc/op   delta
Write1k-8      61.6MB ± 0%    35.7MB ± 0%   -42.02%  (p=0.000 n=10+10)
Write16k-8     27.9MB ± 0%     2.2MB ± 0%   -91.99%  (p=0.000 n=10+10)
Write32k-8     29.9MB ± 0%     1.1MB ± 0%   -96.26%  (p=0.000 n=10+9)
Write128k-8    29.2MB ± 0%     0.3MB ± 0%   -98.87%  (p=0.000 n=10+9)
Write512k-8    29.0MB ± 0%     0.1MB ± 0%   -99.55%  (p=0.000 n=10+10)
Write1MiB-8    28.9MB ± 0%     0.1MB ± 0%   -99.66%  (p=0.000 n=10+10)

name         old allocs/op  new allocs/op  delta
Write1k-8       92.2k ± 0%     82.0k ± 0%   -11.12%  (p=0.000 n=9+10)
Write16k-8      5.81k ± 0%     5.16k ± 0%   -11.23%  (p=0.000 n=10+9)
Write32k-8      2.93k ± 0%     2.60k ± 0%   -11.35%  (p=0.000 n=10+10)
Write128k-8     2.45k ± 0%     2.12k ± 0%   -13.56%  (p=0.002 n=8+10)
Write512k-8     2.33k ± 0%     2.00k ± 0%   -14.30%  (p=0.000 n=9+9)
Write1MiB-8     2.31k ± 0%     1.98k ± 0%   -14.42%  (p=0.000 n=6+9)
```

And against this PR (new = yours):

```
name         old time/op    new time/op    delta
Write1k-8       326ms ± 7%     311ms ± 6%   -4.54%  (p=0.043 n=10+10)
Write16k-8     35.5ms ± 7%    32.7ms ± 4%   -7.93%  (p=0.000 n=10+10)
Write32k-8     19.8ms ± 5%    17.6ms ± 5%  -10.89%  (p=0.000 n=10+10)
Write128k-8    14.9ms ± 4%    14.7ms ± 8%     ~     (p=0.971 n=10+10)
Write512k-8    10.4ms ± 4%    10.2ms ±10%     ~     (p=0.912 n=10+10)
Write1MiB-8    9.33ms ± 1%    9.23ms ± 3%     ~     (p=0.182 n=9+10)
Write4MiB-8    8.86ms ± 7%    8.81ms ± 4%     ~     (p=0.481 n=10+10)

name         old speed      new speed      delta
Write1k-8    32.3MB/s ± 7%  33.8MB/s ± 6%   +4.68%  (p=0.045 n=10+10)
Write16k-8    296MB/s ± 7%   321MB/s ± 4%   +8.53%  (p=0.000 n=10+10)
Write32k-8    530MB/s ± 5%   595MB/s ± 5%  +12.20%  (p=0.000 n=10+10)
Write128k-8   702MB/s ± 4%   715MB/s ± 8%     ~     (p=0.971 n=10+10)
Write512k-8  1.01GB/s ± 4%  1.03GB/s ±10%     ~     (p=0.912 n=10+10)
Write1MiB-8  1.12GB/s ± 1%  1.14GB/s ± 3%     ~     (p=0.182 n=9+10)
Write4MiB-8  1.17GB/s ± 5%  1.19GB/s ± 4%     ~     (p=0.243 n=9+10)

name         old alloc/op   new alloc/op   delta
Write1k-8      35.7MB ± 0%    35.7MB ± 0%   -0.01%  (p=0.000 n=9+10)
Write16k-8     2.24MB ± 0%    2.24MB ± 0%   -0.03%  (p=0.000 n=10+10)
Write32k-8     1.12MB ± 0%    1.12MB ± 0%   -0.05%  (p=0.000 n=10+9)
Write128k-8     330kB ± 0%     330kB ± 0%   -0.17%  (p=0.000 n=10+9)
Write512k-8     132kB ± 0%     131kB ± 0%   -0.35%  (p=0.000 n=10+10)
Write1MiB-8    98.6kB ± 0%    98.1kB ± 0%   -0.49%  (p=0.000 n=10+10)
Write4MiB-8    72.1kB ± 0%    71.6kB ± 0%   -0.67%  (p=0.000 n=10+10)

name         old allocs/op  new allocs/op  delta
Write1k-8       82.0k ± 0%     82.0k ± 0%   -0.03%  (p=0.000 n=10+10)
Write16k-8      5.17k ± 0%     5.16k ± 0%   -0.18%  (p=0.000 n=10+9)
Write32k-8      2.60k ± 0%     2.60k ± 0%   -0.35%  (p=0.000 n=9+10)
Write128k-8     2.12k ± 0%     2.12k ± 0%   -0.40%  (p=0.000 n=10+10)
Write512k-8     2.00k ± 0%     2.00k ± 0%   -0.45%  (p=0.000 n=9+9)
Write1MiB-8     1.99k ± 0%     1.98k ± 0%   -0.45%  (p=0.000 n=10+9)
Write4MiB-8     1.97k ± 0%     1.96k ± 0%   -0.46%  (p=0.000 n=10+10)
```

That's quite a bit better for small writes!