@greatroar 2020-10-29
Here's new=113f5e9b80e37152e3412ee9a2d6fbd8c61e9b37 against old=2b85c58483d85114825b740918f7fed96cf330cf (the delayed benchmarks won't run on master):

```
name                     old time/op    new time/op    delta
Write4MiBDelay10Msec-8      260ms ± 0%     259ms ± 0%   -0.40%  (p=0.000 n=8+8)
Write4MiBDelay50Msec-8      1.26s ± 0%     1.26s ± 0%   -0.06%  (p=0.000 n=7+9)
Write4MiBDelay150Msec-8     3.76s ± 0%     3.76s ± 0%   -0.05%  (p=0.000 n=8+10)

name                     old speed      new speed      delta
Write4MiBDelay10Msec-8   40.4MB/s ± 0%  40.5MB/s ± 0%   +0.43%  (p=0.000 n=9+8)
Write4MiBDelay50Msec-8   8.32MB/s ± 0%  8.32MB/s ± 0%     ~     (all equal)
Write4MiBDelay150Msec-8  2.79MB/s ± 0%  2.79MB/s ± 0%     ~     (all equal)

name                     old alloc/op   new alloc/op   delta
Write4MiBDelay10Msec-8     42.0MB ± 0%    10.6MB ± 0%  -74.84%  (p=0.000 n=10+9)
Write4MiBDelay50Msec-8     42.0MB ± 0%    10.6MB ± 0%  -74.83%  (p=0.000 n=9+9)
Write4MiBDelay150Msec-8    42.0MB ± 0%    10.6MB ± 0%  -74.83%  (p=0.000 n=10+10)

name                     old allocs/op  new allocs/op  delta
Write4MiBDelay10Msec-8      2.61k ± 0%     2.61k ± 0%   -0.13%  (p=0.000 n=10+8)
Write4MiBDelay50Msec-8      2.62k ± 0%     2.62k ± 0%   -0.09%  (p=0.010 n=9+9)
Write4MiBDelay150Msec-8     2.62k ± 0%     2.62k ± 0%   -0.08%  (p=0.006 n=10+10)
```

With f80745d84ca72104a8f0635ae1b6d6e1a91c600d applied, this becomes (old is now 113f5e9b80e37152e3412ee9a2d6fbd8c61e9b37):

```
name                     old time/op    new time/op     delta
Write4MiBDelay10Msec-8      259ms ± 0%       91ms ± 1%   -64.78%  (p=0.000 n=8+9)
Write4MiBDelay50Msec-8      1.26s ± 0%      0.41s ± 0%   -67.34%  (p=0.000 n=9+10)
Write4MiBDelay150Msec-8     3.76s ± 0%      1.21s ± 0%   -67.76%  (p=0.000 n=10+10)

name                     old speed      new speed       delta
Write4MiBDelay10Msec-8   40.5MB/s ± 0%  115.1MB/s ± 1%  +183.93%  (p=0.000 n=8+9)
Write4MiBDelay50Msec-8   8.32MB/s ± 0%  25.49MB/s ± 0%  +206.35%  (p=0.000 n=8+10)
Write4MiBDelay150Msec-8  2.79MB/s ± 0%   8.65MB/s ± 0%  +210.04%  (p=0.000 n=10+7)

name                     old alloc/op   new alloc/op    delta
Write4MiBDelay10Msec-8     10.6MB ± 0%     10.6MB ± 0%    +0.18%  (p=0.000 n=9+9)
Write4MiBDelay50Msec-8     10.6MB ± 0%     10.6MB ± 0%    +0.16%  (p=0.000 n=9+9)
Write4MiBDelay150Msec-8    10.6MB ± 0%     10.6MB ± 0%    +0.19%  (p=0.000 n=10+10)

name                     old allocs/op  new allocs/op   delta
Write4MiBDelay10Msec-8      2.61k ± 0%      3.00k ± 0%   +15.26%  (p=0.000 n=8+9)
Write4MiBDelay50Msec-8      2.62k ± 0%      3.01k ± 1%   +14.96%  (p=0.000 n=9+9)
Write4MiBDelay150Msec-8     2.62k ± 0%      3.02k ± 1%   +15.33%  (p=0.000 n=10+10)

```

If you want to try this too, here's what I did:

```
$ go install golang.org/x/perf/cmd/benchstat
$ TMPDIR=/dev/shm go test -run=- -integration -bench='Write4MiBDelay' -benchmem -count=10 | tee -a log0
# similarly for log1, after checking out a different commit
$ benchstat log0 log1
```

Benchstat can also take more than two logs as input, but then it can't do significance testing.